requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pymodbus==3.6.8
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
yarl==1.9.4
